Add bundle to install just the Phoronix-test-suite
Add Bundle to support bare Phoronix-test-suite(pts). This is useful for
for users that want to install only the test suite and solve the
dependencies of their test cases depending upon their needs.

The pts can be installed for system-wide usage or run locally. The only
hard dependency on the Phoronix test suite is having command-line
support for PHP (PHP 5.3+) installed. A complete PHP stack (e.g. with
web server) is not needed.

If users want to run general stuff the other bundles solve all their
dependencies:

os-testsuite-phoronix
os-testsuite-phoronix-desktop
os-testsuite-phoronix-server

Signed-off-by: Victor Rodriguez <victor.rodriguez.bahena@intel.com>