Add temporary deprecated python2 bundle
This bundle is intended to be an emergency release in case something
critical turns out to need python2 still. It is not intended to be
anything more than a temporary last resort. The bundle will be removed
without notice shortly assuming most of the rest of the world is doing
just fine without python2.